Create python
This code creates an FIR filter with a cutoff frequency of 100 Hz and a length of 50 points using the firwin function from the SciPy library. It then displays the frequency response of the filter and applies it to an input signal using the lfilter function. Note that input_signal should be your actual input signal that you want to filter.